+++ ehpcinstant-20230701/core/Models/ListJobExecutorsResponseBody.cs
@@ -55,6 +55,10 @@ namespace AlibabaCloud.SDK.EhpcInstant20230701.Models
             [Validation(Required=false)]
             public int? ArrayIndex { get; set; }
 
+            [NameInMap("BlockDuration")]
+            [Validation(Required=false)]
+            public int? BlockDuration { get; set; }
+
             /// <summary>
             /// <b>Example:</b>
             /// <para>2024-02-20 10:04:10</para>
@@ -90,6 +94,10 @@ namespace AlibabaCloud.SDK.EhpcInstant20230701.Models
             [NameInMap("IpAddress")]
             [Validation(Required=false)]
             public List<string> IpAddress { get; set; }
+
+            [NameInMap("Preemptible")]
+            [Validation(Required=false)]
+            public bool? Preemptible { get; set; }
 
             [NameInMap("StartTime")]
             [Validation(Required=false)]